Changes for this commit include fixing sign in issues between the view controllers (some using FIRAuth()? etc and others use FIRAuth! etc and they have different results/return values).
Also included in this commit was the integration of pop up messages that will appear on (currently just) the home screen (eventually any and every screen in the app) and the changes to the back end that said pop up messages will make. A driver accepting a rider and thus giving the rider their UID in the driver_uid field on firebase.